Import rules

+++ Source/Our.Umbraco.HideProperties/HidePropertiesContext.cs
@@ -1,13 +1,15 @@
 ﻿namespace Our.Umbraco.HideProperties
 {
     using System;
+    using System.Collections.Generic;
     using System.IO;
 
     using global::Umbraco.Core.IO;
     using global::Umbraco.Core.Logging;
 
     using Newtonsoft.Json;
-
+    using Our.Umbraco.HideProperties.ContractResolver;
+    using Our.Umbraco.HideProperties.Models.Pocos;
     using Our.Umbraco.HideProperties.Models.Repositories;
 
     /// <summary>
@@ -21,11 +23,20 @@
         private static HidePropertiesContext instance;
 
         /// <summary>
+        /// Custom contract resolver
+        /// </summary>
+        private readonly IgnorableSerializerContractResolver contractResolver;
+
+        /// <summary>
         /// Prevents a default instance of the <see cref="HidePropertiesContext"/> class from being created.
         /// </summary>
         private HidePropertiesContext()
         {
-            this.Configuration = HidePropertiesConfig.Current;            
+            this.Configuration = HidePropertiesConfig.Current;
+
+            // Ignore rule.id
+            this.contractResolver = new IgnorableSerializerContractResolver().Ignore<Rule>(rule => rule.Id);
+
             instance = this;
         }
 
@@ -44,23 +55,91 @@
         {
             try
             {
-                var rulesFile = IOHelper.MapPath(Path.Combine(SystemDirectories.Config, "hideProperties.rules.js"));
-
-                if (File.Exists(rulesFile))
+                if (this.Configuration.IsExportEnabled)
                 {
-                    File.Delete(rulesFile);
-                }
+                    var rulesFile = IOHelper.MapPath(Path.Combine(SystemDirectories.Config, "hideProperties.rules.js"));
 
-                using (var file = File.CreateText(rulesFile))
-                {
-                    var serializer = new JsonSerializer { Formatting = Formatting.Indented };
+                    if (File.Exists(rulesFile))
+                    {
+                        File.Delete(rulesFile);
+                    }
 
-                    serializer.Serialize(file, RuleRepository.Current.Get());
+                    using (var file = File.CreateText(rulesFile))
+                    {
+                        var serializer = new JsonSerializer { Formatting = Formatting.Indented, ContractResolver = this.contractResolver };
+
+                        serializer.Serialize(file, RuleRepository.Current.Get());
+                    }
                 }
             }
             catch (Exception ex)
             {
                 LogHelper.Warn<Startup>("Unable to save rules to disk: {0}", () => ex.ToString());
+            }
+        }
+
+        public void ImportRules()
+        {
+            try
+            {
+                if (this.Configuration.IsImportEnabled)
+                {
+                    var rulesFile = IOHelper.MapPath(Path.Combine(SystemDirectories.Config, "hideProperties.rules.js"));
+
+                    if (File.Exists(rulesFile))
+                    {
+                        var serializer = new JsonSerializer { Formatting = Formatting.Indented, ContractResolver = this.contractResolver };
+
+                        using (var file = new StreamReader(rulesFile))
+                        {
+                            using (var reader = new JsonTextReader(file))
+                            {
+                                var rules = serializer.Deserialize<IEnumerable<Rule>>(reader);
+
+                                using (var transaction = RuleRepository.Current.Database.GetTransaction())
+                                {
+                                    foreach (var rule in rules)
+                                    {
+                                        var existingRule = RuleRepository.Current.GetSingleBy(x => x.Key.Equals(rule.Key));
+
+                                        if (existingRule != null)
+                                        {
+                                            if (this.Configuration.UpdateRuleAtImport)
+                                            {
+                                                rule.Id = existingRule.Id;
+
+                                                RuleRepository.Current.Save(rule);
+                                            }
+                                            else
+                                            {
+                                                if (this.Configuration.DeleteRuleAtImport && rule.IsDeleted)
+                                                {
+                                                    existingRule.IsDeleted = true;
+
+                                                    RuleRepository.Current.Save(existingRule);
+                                                }
+                                            }
+                                        }
+                                        else
+                                        {
+                                            RuleRepository.Current.Add(rule);
+                                        }
+                                    }
+
+                                    transaction.Complete();
+                                }
+                            }
+                        }
+                    }
+                    else
+                    {
+                        LogHelper.Warn<Startup>("Unable to import rules from disk: Missing hideProperties.rules.js file");
+                    }
+                }
+            }
+            catch (Exception ex)
+            {
+                LogHelper.Warn<Startup>("Unable to import rules from disk: {0}", () => ex.ToString());
             }
         }
     }
